Fix all clippy warnings: zero warnings across lib + 14 examples
Library (src/) — 8 warnings fixed:

  src/led.rs:
  - from_over_into: Replace impl Into<Led> with impl From<Led> in ctor! macro
    (clippy::from_over_into, 4 instances from macro expansion)

  src/sdram.rs:
  - type_complexity: Extract 55-element pin tuple return type into SdramPinsTuple
    type alias for split_sdram_pins() (clippy::type_complexity)
  - manual_is_multiple_of: Replace `n % m == 0` with `n.is_multiple_of(m)` in
    as_slice_mut() and subslice_mut() alignment checks (clippy::manual_is_multiple_of, 3 sites)

  src/lcd.rs (framebuffer feature):
  - needless_borrow: Remove unnecessary `&mut` reborrow in back_buffer() and
    front_buffer() return expressions (clippy::needless_borrow, 2 sites)

Examples — 27 warnings fixed:

  test_all.rs (11 warnings):
  - unnecessary_mut_passed: Remove &mut from adc::Temperature and adc::Vref
    (clippy::unnecessary_mut_passed, 2 sites)
  - unused_enumerate_index: Remove .enumerate() when index unused in SDRAM
    verification loops (clippy::unused_enumerate_index, 4 sites)
  - single_match: Convert single-pattern match on I2C touch read to if let
    (clippy::single_match)
  - manual_range_contains: Use (lo..=hi).contains(&v) for timer tolerance and
    touch coordinate bounds checks (clippy::manual_range_contains, 3 sites)

  hw_diag.rs (12 warnings):
  - needless_range_loop: Replace `for i in range { &results[i] }` with direct
    slice iteration `for result in &results[range]` in all 7 render sections
    (clippy::needless_range_loop)
  - manual_range_contains: Use range contains for timer checks and touch bounds
    (clippy::manual_range_contains, 3 sites)
  - single_match: Convert single-pattern match on I2C touch read to if let
    (clippy::single_match)

  test_timers.rs (6 warnings):
  - manual_range_contains: Use range contains for all 4 timer tolerance checks
    (clippy::manual_range_contains)
  - let_unit_value: Remove `let _ =` for unit-returning pwm.set_period() calls
    (clippy::let_unit_value, 2 sites)

  test_lcd.rs (3 warnings):
  - unnecessary_cast: Remove redundant `as u32` casts in rainbow gradient
    calculation where expression is already u32 (clippy::unnecessary_cast, 3 sites)

  test_led.rs (1 warning):
  - explicit_auto_deref: Replace &*leds with &leds for array-to-slice coercion
    (clippy::explicit_auto_deref)

  test_soak.rs (3 warnings):
  - manual_is_multiple_of: Replace `n % m == 0` with `n.is_multiple_of(m)` for
    pattern alternation and logging intervals (clippy::manual_is_multiple_of, 3 sites)

  test_usb.rs (2 warnings):
  - manual_is_multiple_of: Replace polls % 500 with polls.is_multiple_of(500)
    (clippy::manual_is_multiple_of)
  - nonminimal_bool: Replace `!x.is_err()` with `x.is_ok()` for serial poll
    (clippy::nonminimal_bool)

  test_usb_standalone.rs (1 warning):
  - manual_is_multiple_of: Replace poll_count % 1000 with
    poll_count.is_multiple_of(1000) (clippy::manual_is_multiple_of)

  test_touch.rs (2 warnings):
  - manual_range_contains: Use range contains for 3px edge margin filter
    (clippy::manual_range_contains)

All changes are mechanical refactors with zero behavioral change.
Verified: cargo clippy --release --lib [--features framebuffer] and all 14
examples report zero warnings.

+++ examples/test_lcd.rs
@@ -304,9 +304,9 @@ fn main() -> ! {
     let mut hue = 0u32;
     loop {
         for row in 0..HEIGHT {
-            let r = ((hue + row as u32 * 3) % 256) as u32;
-            let g = ((hue + row as u32 * 3 + 85) % 256) as u32;
-            let b = ((hue + row as u32 * 3 + 170) % 256) as u32;
+            let r = (hue + row as u32 * 3) % 256;
+            let g = (hue + row as u32 * 3 + 85) % 256;
+            let b = (hue + row as u32 * 3 + 170) % 256;
             let color = (r << 16) | (g << 8) | b;
             let start = row * WIDTH;
             for pixel in fb.iter_mut().take(WIDTH).skip(start) {
